+++ changelog.md
@@ -1,4 +1,8 @@
-* 05-10-2025 : feat : voice search [PR #14](https://github.com/ShubhamOulkar/weather/pull/14)
+- 05-10-2025 : feat + testing [PR#13](https://github.com/ShubhamOulkar/weather/pull/13)
+    - Add animations
+    - add react query hooks testing
+    
+- 05-10-2025 : feat : voice search [PR #14](https://github.com/ShubhamOulkar/weather/pull/14)
     - Starts listening when the button is clicked.
     - Automatically stops after 10 seconds.
     - Updates the search query live while listening.